Move events class in workflow
