+++ events/code_review_events/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+MONITORING_PERIOD = 7 * 3600
+
 QUEUE_MERCURIAL = "mercurial"
 QUEUE_MONITORING = "monitoring"
 QUEUE_PHABRICATOR_RESULTS = "results"
